+++ src/docs/release-notes.md
@@ -2,6 +2,7 @@
 
 ## Version 1.3-alpha | unreleased
 
+- Added support for search buckets
 - Fixed compiler errors introduced by the latest TypeScript version
 - Replaced deprecated calls to `substr` with `substring`